Make the zephyr-chrome repository a zephyr module
This adds a "zephyr/module.yml" file to the root of the zephyr-chrome
repo, enabling code sharing across projects.  As an example, the
posix-ec project was migrated to the new project format, which
requires no C sources in the project directory.

The common main function now prints a message, as well as the active
copy.

BUG=b:168560968
TEST=run posix-ec app, observe output

Change-Id: I9e68555a4dccf22778882950f349364abaaba1f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2411542
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>