chameleon: enable Zephyr shell
Add the shell subsystem to the Chameleon app, and add a simple
command to turn the PG_CHAMELEON LED off and on.

BUG=b:169245422
TEST=build and flash the app as documented in README.md
Open a terminal on the UART and reset the board.
When the `uart:~$` prompt appears, use the commands "led on" and
"led off" to control the red LED at the bottom edge of the board
near the PCIe connector. Type "led" and you get a usage summary.

Change-Id: I9ce254df0557fb49e8f40e3353b1126d64c9423a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2426447
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>