chameleon: enable the ADCs
* Enable the ADC driver.
* Add a "sysmon" module that owns the ADC devices and the GPIO
for SYSMON_SEL.
* Add terribly-hacky code that reads the ADCs and dumps out
the raw values just to prove the ADCs work.

BUG=b:169366335
TEST=build and flash the app as documented in README.md
Open a terminal on the UART and reset the board.
When the `uart:~$` prompt appears, use the command "read_adcs".
The values printed are for VMON_12V, the two current monitors,
and then SYSMON1:8.

Change-Id: I8f3a59edfaae0620015f234186f7a529445fa08e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2441975
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>