chameleon: update single-line blocks to Zephyr coding style
Update the single-line code blocks to match the Zephyr coding style
https://docs.zephyrproject.org/latest/contribute/index.html#coding-style
which calls for braces on "every if, else, do, while, file and switch
body, even for single-line code blocks." This is also in line with
MISRA coding standards.

This is an exception to the Linux kernel coding style, so add an
option to .checkpatch.conf to prevent checkpatch from warning about
it.

BUG=None
TEST=build the code and ad-hoc test it on the board

Change-Id: Ibe6140bd6af3ba0a616561448e381fdd2df769ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2488500
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Auto-Submit: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>