chameleon: implement system monitor
Change sysmon_init to start a separate task that reads the ADC inputs
and saves the readings internally.
Remove the `read_adcs` command and replace with `sysmon` to print the
values in a friendly format.

BUG=b:169365982
TEST=build and program to the board. In the shell run `sysmon` and
check that the reported values are reasonable. Note that the SOM
values won't be reasonable unless/until the FPGA SOM is installed
and powered up.

Change-Id: I72a4912c60f2fa9d25a0154021abc39dd130cbcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2490656
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>