tests: move base32 test back to EC
chromium:2492527 added support for the Ztest API in the EC tests,
so we don't need a separate (and possibly out-of-sync) copy of the
tests here.
Also remove the HOWTO document, as that information will be updated
and included in the EC documentation.

This is a revert of most (but not all) of chromium:2468630 and a
full revert of chromium:2472757.

BUG=b:168032590
TEST=n/a

Cq-Depend: chromium:2492527
Change-Id: I8d3d6eabc23e83d29754f047d2b5156737e09b02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2495097
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>
Commit-Queue: Yuval Peress <peress@chromium.org>
Auto-Submit: Paul Fagerburg <pfagerburg@chromium.org>