chameleon: refactor GPIO handling
Move the GPIO functions out of the SD card mux module and improve
them so that multiple modules can use GPIO with minimal duplicated
code.
Add documentation for gpio_signal.h to explain why it exists and
how to use it.

BUG=b:171439846
TEST=build the code and run the ad-hoc SD card test (`sd usb` and
then insert an SD card, see it show up as a mounted drive on the
computer), as well as the ad-hoc SYSMON test (`sysmon` and see
that the values make sense, which indicates that the SYSMON_SEL
GPIO is being controlled correctly).

Change-Id: I361f8b0994d9a84f3b73507838a36d13d7267b6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2506034
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>