chameleon: use a timer to trigger sysmon
Instead of using a k_sleep in the sysmon task, use a timer to signal
a semaphore that the sysmon task will take. The timer sets the
cadence for sampling. It is easier to change the timer at the top of
the file than to dig into the sysmon task code.

BUG=None
TEST=Build and run as detailed in the "System Monitor" section
of README.md

Change-Id: I2056e656c4dae9cd1414a74de4f172d365853b25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2535083
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>
Commit-Queue: Yuval Peress <peress@chromium.org>