chameleon: enable video muxes
Add an API for controlling the routing of the HDMI and DP signals to
the FPGA's MGTs and the HDMI receiver IC. Add shell commands for
using the API.

BUG=b:174808617
TEST=compile and run
Use the command `videomux 1|2 hdmi|dp` to select the HDMI or DP
signals for ports 1 or 2. However, since we're not asserting HPD
or decoding any signals, you won't be able to observe any effect
from running these commands.

Change-Id: I3b1d278d85e05f2cee7d12de1a69286582683fa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/zephyr-chrome/+/2585558
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>