lazor: move declaration of battery_cell_type to ln9310.h
ln9310 driver depends on these functions (sadly), so it does not make
sense to be declaring them in board.h.  Specifically, the Zephyr build
won't include Lazor's board.h.

ln9310.h is the best home I can think of for now.  Ideally we could
refactor this later so that the ln9310 driver wasn't depending on
board-specific functions.

BUG=b:183054226
BRANCH=none
TEST=buildall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I7c82d2ed3d3c16d8131238cd236741e31de5b0c7