lazor: Create a new file for USB-C config
With Zephyr we cannot compile board.c so we need to move the USB-C
code into a file that can be built by both Zephyr and ECOS.

Make a start on a new ucbc_config.c file, moving just the
tcpc_alert_event() function, to start.

Update build.mk to separate out the files, keeping them in order. This
will make it easier to read as the list grows.

BUG=b:183296099
BRANCH=none
TEST=make BOARD=trogdor -j4

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I8943a9732408a1d3ce6a80c99d6a7d3565fbd7e4