zephyr: Add support for hardware charge ramping
Allow this option to be selected, turning it in a choice between the
two options. Make hardware ramping the default since it seems to be
preferred.

Also move the requirement for software ramp for ISL9241 to Kconfig. We
cannot 'select' a choice item like PLATFORM_EC_CHARGE_RAMP_SW in
PLATFORM_EC_CHARGER_ISL9241, so use a 'depends on' instead. Update
volteer to enable software ramping since the default has changed.

BUG=b:183296099
BRANCH=none
TEST=build for lazor
build for volteer and check that CONFIG_PLATFORM_EC_CHARGE_RAMP_SW=y

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ic3ffd21cd10aa896a4c6dbe1d98d75776c89aefe