zephyr: Tidy up BATTERY_PRESENT_GPIO
This is not a boolean CONFIG in ECOS but it must be with Zephyr. There
doesn't seem to be any advantage to allowing an arbitrary name here,
since the mapping is controlled in the device tree anyway.

So hard-code it to GPIO_BATT_PRES_ODL and document this in the Kconfig.

Also drop the unecessary logic in config_chip.h since Kconfig handles
ensuring that only one option is set. This makes the code easier to
follow and avoids errors. Fix the error that is there.

BUG=b:183296099
BRANCH=none
TEST=make BOARD=lazor -j4

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I55a9083ad01be2f6b1e9314f73fbbff1442e486c